rename: utils 함수 분류 - 서비스를 위한 유틸리티 함수들: serviceUtils
