+++ src/screens/Components/BigCardList.js
@@ -1,5 +1,5 @@
 import Component from "../Component.js";
-import { createExtendsRelation, getComponentsTemplate } from "../../utils.js";
+import { createExtendsRelation } from "../../utils.js";
 import BannerImage from "./BannerImage.js";
 
 function BigCardList() {}
