+++ src/serviceUtils.js
@@ -1,0 +1,19 @@
+const formatUserCount = (userCount) => {
+  return (userCount / 10000).toFixed(1) + "만명";
+};
+
+const getComponentsTemplate = (components) => {
+  return components?.reduce((tags, component) => {
+    tags += component.template();
+    return tags;
+  }, "");
+};
+
+const updateNodeClasses = (target, className) => {
+  [...target.parentNode.children].forEach((node) => {
+    node.classList.remove(className);
+  });
+  target.classList.add(className);
+};
+
+export { formatUserCount, getComponentsTemplate, updateNodeClasses };
